+++ includes/Plugin/Beaver_Builder/Component.php
@@ -8,7 +8,8 @@
  * @package    Zooey
  * @copyright  WebMan Design, Oliver Juhas
  *
- * @since  1.0.0
+ * @since    1.0.0
+ * @version  1.1.9
  */
 
 namespace WebManDesign\Zooey\Plugin\Beaver_Builder;
@@ -42,7 +43,8 @@ class Component implements Component_Interface {
 	/**
 	 * Upgrade URL.
 	 *
-	 * @since  1.0.0
+	 * @since    1.0.0
+	 * @version  1.1.9
 	 *
 	 * @param  string $url
 	 *
@@ -52,7 +54,7 @@ class Component implements Component_Interface {
 
 		// Output
 
-			return esc_url( add_query_arg( 'fla', '67', (string) $url ) );
+			return trailingslashit( FL_BUILDER_STORE_URL ) . 'fla/67/';
 
 	} // /upgrade_url