+++ patterns/numbers/numbers-05.php
@@ -5,7 +5,8 @@
  * @package    Zooey
  * @copyright  WebMan Design, Oliver Juhas
  *
- * @since  1.0.0
+ * @since    1.0.0
+ * @version  1.1.9
  */
 
 namespace WebManDesign\Zooey\Content;
@@ -34,12 +35,12 @@ Block_Pattern::add_pattern_args( __FILE__, array(
 		<!-- wp:column {"style":{"spacing":{"blockGap":"var:preset|spacing|s"}}} -->
 		<div class="wp-block-column">
 
-			<!-- wp:paragraph {"style":{"typography":{"lineHeight":"1","fontSize":"8em"},"elements":{"link":{"color":{"text":"var:preset|color|secondary"}}},"spacing":{"margin":{"bottom":"var:preset|spacing|l"}}},"textColor":"secondary","fontFamily":"supplemental"} -->
-			<p class="has-secondary-color has-text-color has-link-color has-supplemental-font-family" style="margin-bottom:var(--wp--preset--spacing--l);font-size:8em;line-height:1">01</p>
+			<!-- wp:paragraph {"style":{"typography":{"lineHeight":"1","fontSize":"8em"},"elements":{"link":{"color":{"text":"var:preset|color|secondary"}}},"spacing":{"margin":{"bottom":"var:preset|spacing|m"}}},"textColor":"secondary","fontFamily":"supplemental"} -->
+			<p class="has-secondary-color has-text-color has-link-color has-supplemental-font-family" style="margin-bottom:var(--wp--preset--spacing--m);font-size:8em;line-height:1">01</p>
 			<!-- /wp:paragraph -->
 
-			<!-- wp:heading -->
-			<h2 class="wp-block-heading"><?php Block_Pattern::the_text( 'title/s' ); ?></h2>
+			<!-- wp:heading {"style":{"spacing":{"margin":{"top":"0"}}}} -->
+			<h2 class="wp-block-heading" style="margin-top:0"><?php Block_Pattern::the_text( 'title/s' ); ?></h2>
 			<!-- /wp:heading -->
 
 			<!-- wp:paragraph -->
@@ -52,12 +53,12 @@ Block_Pattern::add_pattern_args( __FILE__, array(
 		<!-- wp:column {"style":{"spacing":{"blockGap":"var:preset|spacing|s"}}} -->
 		<div class="wp-block-column">
 
-			<!-- wp:paragraph {"style":{"typography":{"lineHeight":"1","fontSize":"8em"},"elements":{"link":{"color":{"text":"var:preset|color|secondary"}}},"spacing":{"margin":{"bottom":"var:preset|spacing|l"}}},"textColor":"secondary","fontFamily":"supplemental"} -->
-			<p class="has-secondary-color has-text-color has-link-color has-supplemental-font-family" style="margin-bottom:var(--wp--preset--spacing--l);font-size:8em;line-height:1">02</p>
+			<!-- wp:paragraph {"style":{"typography":{"lineHeight":"1","fontSize":"8em"},"elements":{"link":{"color":{"text":"var:preset|color|secondary"}}},"spacing":{"margin":{"bottom":"var:preset|spacing|m"}}},"textColor":"secondary","fontFamily":"supplemental"} -->
+			<p class="has-secondary-color has-text-color has-link-color has-supplemental-font-family" style="margin-bottom:var(--wp--preset--spacing--m);font-size:8em;line-height:1">02</p>
 			<!-- /wp:paragraph -->
 
-			<!-- wp:heading -->
-			<h2 class="wp-block-heading"><?php Block_Pattern::the_text( 'title/s' ); ?></h2>
+			<!-- wp:heading {"style":{"spacing":{"margin":{"top":"0"}}}} -->
+			<h2 class="wp-block-heading" style="margin-top:0"><?php Block_Pattern::the_text( 'title/s' ); ?></h2>
 			<!-- /wp:heading -->
 
 			<!-- wp:paragraph -->
@@ -70,12 +71,12 @@ Block_Pattern::add_pattern_args( __FILE__, array(
 		<!-- wp:column {"style":{"spacing":{"blockGap":"var:preset|spacing|s"}}} -->
 		<div class="wp-block-column">
 
-			<!-- wp:paragraph {"style":{"typography":{"lineHeight":"1","fontSize":"8em"},"elements":{"link":{"color":{"text":"var:preset|color|secondary"}}},"spacing":{"margin":{"bottom":"var:preset|spacing|l"}}},"textColor":"secondary","fontFamily":"supplemental"} -->
-			<p class="has-secondary-color has-text-color has-link-color has-supplemental-font-family" style="margin-bottom:var(--wp--preset--spacing--l);font-size:8em;line-height:1">03</p>
+			<!-- wp:paragraph {"style":{"typography":{"lineHeight":"1","fontSize":"8em"},"elements":{"link":{"color":{"text":"var:preset|color|secondary"}}},"spacing":{"margin":{"bottom":"var:preset|spacing|m"}}},"textColor":"secondary","fontFamily":"supplemental"} -->
+			<p class="has-secondary-color has-text-color has-link-color has-supplemental-font-family" style="margin-bottom:var(--wp--preset--spacing--m);font-size:8em;line-height:1">03</p>
 			<!-- /wp:paragraph -->
 
-			<!-- wp:heading -->
-			<h2 class="wp-block-heading"><?php Block_Pattern::the_text( 'title/s' ); ?></h2>
+			<!-- wp:heading {"style":{"spacing":{"margin":{"top":"0"}}}} -->
+			<h2 class="wp-block-heading" style="margin-top:0"><?php Block_Pattern::the_text( 'title/s' ); ?></h2>
 			<!-- /wp:heading -->
 
 			<!-- wp:paragraph -->